Update ScraperApp.java (Now with timeline restriction capabilities)
The exe in App folder is not updated, but if you compile this code into an exe it will be the latest edition, my next update is timeline span restrictions.

This update was restricting of timeline with stricter date retention rules. Only the following dates are retained in the CompiledSortedTimeline.txt file when the "Restrict Timeline" is toggled, instances of sentences or paragraphs (depending on dropdown selection) that contain a 4 digit year (the first one it finds). This String is only kept in the compiled sorted timeline text file if any of the 4 words to the left of the year are "on" (ignoring case), or if any of the 2 words to the left of it "in" (ignoring case), or if either of the 2 words to the right of it are any form of "AD"/"BC"/"BCE". 

The logic behind this is that some years may be confused with descriptions of quantities of items (for example: "1000 barrels of oil." Is a sentence that would be scraped under the non-strict rules) and to combat this confusion, we add logical restrictions/checks to attempt to limit the amount of non-year sentences/paragraphs collected.

years != 4 digits long aren't included, still (means that 999 BC - 999 AD and < 9999 BC and > 9999 AD are years are that never collected. Also things like "21st century" aren't collected).