Update ScraperApp.java (Now with date range limiting capabilities)
Limit ranges that show up in the compiled sorted timeline file (only this one not the respective document file) from bce to ad dates.

If a user only cares about things that happened from 1900-2000 then they should be able to customize the output in that way without having to sift through a bunch of fluff before 1900 and after 2000.

Next update will be the ability for typos in the links to still find the closest match on wiki. The goal behind this is so that instead of users having to collect many links in the first place, they can just have a general knowledge of relevant events and type them into the links file without having to worry about typos. For example a person may want to research the different species of Dinosaur, so that user will type in the links file "Dinosaurs\nTryceratopes\nT rex\n etc..." I want this user to not have to go and find the exact way to spell these things but rather for the app to automatically find the most likely desired page. Things like "T rex" and "Dinosaurs" already automatically autocorrect and connect to the wiki page's actual address (/Tyrannosaurus and /Dinosaur), but Tryceratopes doesn't automatically connect to the /Triceratops page, so I want the app to automatically select the top result when quering wiki for "Tryceratopes" which would connect us to the desired page Triceratops.

This described update above will be developed after I release a video on the Uri Mittelman Youtube channel (https://www.youtube.com/@UriMittelman) and a post on the Mittelman Journal blog (https://mittelmanjournal.github.io/webjournal/index.html) explaining how to use the tool at its current iteration.