Update ScraperApp.java to now have file path-free capabilities
I noticed that it is a hassle to have to input the file path to the List of Links and List of keywords file, to combat this I introduced the ability to input a comma separated list of values into both the links and keywords input boxes. Users retain the ability to put a file path in the textboxes. The output file's absolute path is still required.

BUGFIX: when year range is empty and sorted timeline is included, the sorted timeline was empty because of a bug, forgot exactly what I did to fix, I think I added a conditional check before  limiting the timeline range to see if it is empty, if so then don't handle any range limitations, if it contains a year then it will look.

Users now also have the ability to not input a full year range, if they only input 1 year (1000 BC or 1900) then that will act as the lower bound while the upper bound is any year ahead of that date. Ability to input range (1000 BC - 2000) is still retained.

Next Updates in dev:
1. Autocorrecting of typos in links to the expected page (with an alert that the scraped page was autocorrected)

2. Instead of writing to .txt (notepad) files write to .docx files (word). In these file types you can create a hyper link to allow users to ctrl click them and find the exact location of the scraped sentence/paragraph for more context in the actual wikipage.